requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.30.0
pillow>=10.0.0
pillow-heif>=0.13.0

# Aesthetic scoring
simple-aesthetics-predictor>=0.1.0

# Similarity computation
scikit-learn>=1.3.0
numpy>=1.24.0

# Progress and utilities
tqdm>=4.65.0
pandas>=2.0.0

# Ollama integration (for composition analysis)
requests>=2.31.0
